map zoom

+++ app/src/main/java/com/roadway/capslabs/roadway_chat/models/RatingVote.java
@@ -12,7 +12,7 @@ public class RatingVote {
         this.text = text;
     }
 
-    public float getStars() {
+    public int getStars() {
         return stars;
     }
